Remove handler interceptor in favour of new body advice
