+++ src/main/kotlin/com/chillibits/coronaaid/jackson/XmlDtdBodyAdvice.kt
@@ -1,0 +1,73 @@
+package com.chillibits.coronaaid.jackson
+
+import com.chillibits.coronaaid.shared.XmlDtdUrl
+import org.springframework.core.MethodParameter
+import org.springframework.http.MediaType
+import org.springframework.http.converter.HttpMessageConverter
+import org.springframework.http.server.ServerHttpRequest
+import org.springframework.http.server.ServerHttpResponse
+import org.springframework.http.server.ServletServerHttpResponse
+import org.springframework.web.bind.annotation.ControllerAdvice
+import org.springframework.web.method.HandlerMethod
+import org.springframework.web.servlet.mvc.method.annotation.ResponseBodyAdvice
+import java.lang.reflect.Method
+import javax.servlet.http.HttpServletResponse
+
+@ControllerAdvice
+class XmlDtdBodyAdvice: ResponseBodyAdvice<Any> {
+
+    private val reflectionCache: MutableMap<Method, String?> = mutableMapOf()
+
+    override fun supports(method: MethodParameter, converter: Class<out HttpMessageConverter<*>>): Boolean {
+        return true
+    }
+
+    override fun beforeBodyWrite(
+            obj: Any?,
+            method: MethodParameter,
+            mediaType: MediaType,
+            converter: Class<out HttpMessageConverter<*>>,
+            serverRequest: ServerHttpRequest, serverResponse: ServerHttpResponse): Any? {
+        if(serverResponse is ServletServerHttpResponse && mediaType == MediaType.APPLICATION_XML) {
+            val servletResponse = serverResponse.servletResponse
+
+            if(servletResponse.status == 200 || servletResponse.status == 201) {
+                handleMethod(serverResponse, method)
+            }
+        }
+
+        return obj
+    }
+
+    private fun handleMethod(response: ServletServerHttpResponse, handler: MethodParameter): Int {
+        if(reflectionCache.containsKey(handler.method)) {
+            return writeToServlet(response.servletResponse, reflectionCache[handler.method])
+        }
+
+        val url = handler.getMethodAnnotation(XmlDtdUrl::class.java)?.url
+        var combined = XML_HEADER
+
+        if(!url.isNullOrEmpty()) {
+            combined += XML_DTD_TEMPLATE.replace("%DTD_URL%", url)
+        }
+
+        reflectionCache[handler.method!!] = combined
+        return writeToServlet(response.servletResponse, combined)
+    }
+
+    private fun writeToServlet(response : HttpServletResponse, data: String?): Int {
+        if(data != null) {
+            val byteArray = data.toByteArray()
+            response.outputStream.write(byteArray)
+
+            return byteArray.size
+        }
+
+        return 0
+    }
+
+    companion object {
+        const val XML_HEADER = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>"
+        const val XML_DTD_TEMPLATE = "<!DOCTYPE Set SYSTEM \"%DTD_URL%\">"
+    }
+}
